Add the `useRbfTransactionData` hook
